Restructure OpenMP code, test nested parallelism. (#208)
Commit a restructuring of the QV OpenMP code. The major difference is
the removal of the shared data code, which was only used to house a
pthread_barrier_t. My testing shows that the use of pthread_barrier_t in
OpenMP regions may be unreliable.

Also commit a test of nested parallelism and how that might look using
QV.

Signed-off-by: Samuel K. Gutierrez <samuel@lanl.gov>