Add first cut of Pthread group split support.
Please note that this is a work in progress, so a lot of cleanup and
fixes are needed.

This adds group splitting capability to Pthreads. That said, there are
still races and memory leaks that I'm working out. In particular, there
appears to be a problem in qvi_hwsplit_coll::split() that results in
inconsistent colorp values (and perhaps other things).

Signed-off-by: Samuel K. Gutierrez <samuel@lanl.gov>