Checkpoint large rework of how threads are dealt with.
Checkpoint code that allows for efficient reuse of thread scopes that
transparently work across Pthreads and OpenMP (assuming a Pthread
back-end).

This also provides a very close version of an example of what we are
presenting in some upcoming work.

I still need to rename and rework some code. This will come soon.

Signed-off-by: Samuel K. Gutierrez <samuel@lanl.gov>