Add initial cut of automatic daemon startup for MPI jobs.
This commit adds support for automatic quo-vadisd startup across the
nodes in a given job.

If QV_PORT is set, then we try to use that port across the nodes. If
unset, then we use a default port for the daemon. For automatic startup
to work, quo-vadisd must be in PATH. Manual daemon startup is still
supported.

Signed-off-by: Samuel K. Gutierrez <samuel@lanl.gov>